refactor: move ReadDat.js -> _shared/read-dat.js
