+++ packages/compile/src/index.js
@@ -1,8 +1,8 @@
 // Copyright (c) 2022 Climate Interactive / New Venture Fund
 
+export { canonicalName } from './_shared/helpers.js'
+export { readDat } from './_shared/read-dat.js'
 export { generateCode } from './CodeGen.js'
 export { parseAndGenerate, printNames } from './Entry.js'
-export { canonicalName } from './_shared/helpers.js'
 export { parseModel } from './Parser.js'
 export { preprocessModel } from './preprocess/preprocessor.js'
-export { readDat } from './ReadDat.js'
